Update importeTotal y existencia Prod

+++ src/main/java/com/example/entrenamiento/model/Venta.java
@@ -20,15 +20,14 @@ public class Venta {
     @NotNull
     private Date fecha;
 
-    private float importe;
+    private double importe;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade=CascadeType.ALL)
     @JoinColumn(name = "idcliente", nullable = false)
     private Cliente cliente;
 
     @OneToMany(mappedBy = "venta", cascade = CascadeType.ALL)
-    //@JoinColumn(name = "idventa", nullable = false)
-        private List<DetalleVenta> detalles=new ArrayList<>();
+    private List<DetalleVenta> detalles=new ArrayList<>();
 
     @Override
     public String toString() {
